PP-5893-Add-MDC-Logging-Filter
Description:
- Renames method to add paymentId, mandateId and refundId to MDC context
- Adds helper method for this
